Condiciones creadas

+++ Server/src/model/CamaUTI.js
@@ -45,15 +45,53 @@ class CamaUTI {
 
     //Eliminar Cama UTI 
     async eliminarCama (hospitalID) {
+        const hospitalRef = db.collection("Hospitales").doc(hospitalID)
+        
+        const camaDoc = await db.collection("Hospitales").doc(hospitalID).collection("CamasUTI").doc(this.camaID).get()
+
+        if (!camaDoc.exists) {
+            throw new Error(`No existe la cama ${this.camaID}.`)
+        }
+
+        const cama = camaDoc.data()
+
+        if (cama.estado) {
+            return false
+        }
+        
         await db.collection("Hospitales").doc(hospitalID).collection("CamasUTI").doc(this.camaID).delete()
+
+        const hospitalData = (await hospitalRef.get()).data()
+        await hospitalRef.update({
+            cantidadCamas: hospitalData.cantidadCamas - 1
+        })
 
         return true
     }
 
     //Ocupar Cama UTI
     async ocuparCama (hospitalID) {
+        const hospitalRef = db.collection("Hospitales").doc(hospitalID)
+
+        const camaDoc = await db.collection("Hospitales").doc(hospitalID).collection("CamasUTI").doc(this.camaID).get()
+
+        if (!camaDoc.exists) {
+            throw new Error(`No existe la cama ${this.camaID}.`)
+        }
+
+        const cama = camaDoc.data()
+
+        if (cama.estado) {
+            return false
+        }
+        
         await db.collection("Hospitales").doc(hospitalID).collection("CamasUTI").doc(this.camaID).update({
             estado: true
+        })
+
+        const hospitalData = (await hospitalRef.get()).data()
+        await hospitalRef.update({
+            cantidadPacientes: hospitalData.cantidadPacientes + 1
         })
 
         this.estado = true
@@ -63,8 +101,27 @@ class CamaUTI {
 
     //Desocupar Cama UTI
     async desocuparCama (hospitalID) {
+        const hospitalRef = db.collection("Hospitales").doc(hospitalID)
+
+        const camaDoc = await db.collection("Hospitales").doc(hospitalID).collection("CamasUTI").doc(this.camaID).get()
+
+        if (!camaDoc.exists) {
+            throw new Error(`No existe la cama ${this.camaID}.`)
+        }
+
+        const cama = camaDoc.data()
+
+        if (!cama.estado) {
+            return false
+        }
+        
         await db.collection("Hospitales").doc(hospitalID).collection("CamasUTI").doc(this.camaID).update({
             estado: false
+        })
+
+        const hospitalData = (await hospitalRef.get()).data()
+        await hospitalRef.update({
+            cantidadPacientes: hospitalData.cantidadPacientes - 1
         })
 
         this.estado = false
